Moved the logic from the darkmode component to the app, passed the mode prop through to the various components that need it, commented out the switch as there is a small issue with it persisting properly
